feat: Add support for generating numbers with floating point precision
The `generate_numbers` function now accepts an additional parameter `etype` to specify whether the generated numbers should be of type 'float' or 'int'. If `etype` is set to 'float', the selected elements will be rounded to the specified precision using random.uniform(). If `etype` is set to 'int', the selected elements will be converted to integers. This enhancement allows for more flexibility in generating numbers with different data types.